* dlip/master:
  Update gitconfig prompt
  Install powerlevel9k zsh theme
  Only disable proxy if necessary

+++ .bash/inc/proxy.sh
@@ -29,12 +29,15 @@ proxy_enable() {
 }
 
 proxy_disable() {
-    rm -f $HOME/.proxy_enable
     unset NO_PROXY ALL_PROXY HTTP_PROXY HTTPS_PROXY no_proxy all_proxy http_proxy https_proxy
-    if [ -n "$BASH_VERSION" ]; then
-        disable-proxy
-    fi
-    if [ "$(uname)" = "Darwin" ]; then
-        networksetup -switchtolocation "No Proxy" > /dev/null
+
+    if [ -e $HOME/.proxy_enable ]; then
+        rm -f $HOME/.proxy_enable
+        if [ -n "$BASH_VERSION" ]; then
+            disable-proxy
+        fi
+        if [ "$(uname)" = "Darwin" ]; then
+            networksetup -switchtolocation "No Proxy" > /dev/null
+        fi
     fi
 }
